+++ src/main/outro_loop.cpp
@@ -448,15 +448,15 @@ void SetupCredits()
     AddCredit(g_gameInfo.creditsHomePage);
 
     CreditOffsetY = (XRender::TargetH + 40);
-    CreditTotalHeight = 32.0;
+    CreditTotalHeight = 32;
 
     for(A = 1; A <= numCredits; A++)
     {
         auto &cr = Credit[A];
         cr.Location.Width = SuperTextPixLen(GetS(cr.Text), g_gameInfo.creditsFont);
         cr.Location.Height = 16;
-        cr.Location.X = (double(XRender::TargetW) / 2) - (cr.Location.Width / 2.0);
+        cr.Location.X = (XRender::TargetW / 2) - (cr.Location.Width / 2);
         cr.Location.Y = 32 * A;
-        CreditTotalHeight += 32.0;
+        CreditTotalHeight += 32;
     }
 }
